many changes

+++ merge_partition_parts.py
@@ -3,10 +3,10 @@
 import sys, getopt
 import glob,os
 
-help_message = 'usage example: python merge_partition_parts.py -r 1 -i /project/home/cluster_vectors/ -o read_partitions/'
+help_message = 'usage example: python merge_partition_parts.py -i /project/home/cluster_vectors/ -o read_partitions/'
 if __name__ == "__main__":
 	try:
-		opts, args = getopt.getopt(sys.argv[1:],'hr:i:o:',["inputdir="])
+		opts, args = getopt.getopt(sys.argv[1:],'hi:o:',["inputdir="])
 	except:
 		print help_message
 		sys.exit(2)
@@ -14,8 +14,6 @@ if __name__ == "__main__":
 		if opt in ('-h','--help'):
 			print help_message
 			sys.exit()
-		elif opt in ('-r',"--filerank"):
-			fr = int(arg)-1
 		elif opt in ('-i','--inputdir'):
 			inputdir = arg
 			if inputdir[-1] != '/':
@@ -25,10 +23,12 @@ if __name__ == "__main__":
 			if outputdir[-1] != '/':
 				outputdir += '/'
 	FP = glob.glob(os.path.join(inputdir,'*.fastq.*'))
-	FP = list(set([fp[fp.rfind('/')+1:fp.index('.fastq')] for fp in FP]))
-	FP.sort()
-	group = FP[fr]
-	gp = glob.glob(os.path.join(inputdir,group+'.fastq.*'))
-	os.system('cat '+' '.join(gp)+' > '+outputdir+group+'.fastq')
-	os.system('touch %s.empty' % (gp[0]))
-	os.system('rm '+' '.join(gp))
+	FPl = list(set([fp[fp.rfind('/')+1:fp.index('.fastq')] for fp in FP]))
+	FPl.sort()
+	for group in FPl:
+		gp = [fp for fp in FP if inputdir+group+'.fastq' == fp[:fp.rfind('.')]]
+		gp = [fp for fp in gp if '.empty' not in fp]
+		if len(gp) > 0:
+			os.system('cat '+' '.join(gp)+' > '+outputdir+group+'.fastq')
+			os.system('touch %s.empty' % (gp[0]))
+			os.system('rm '+' '.join(gp))
